requirements.txt
pillow==10.2.0
matplotlib==3.8.3
streamlit==1.31.1
pytesseract==0.3.10